chore: rearrange utils
